Fix and updates to setup script, thanks @SamSaffron
- Add swap if it doesn't exist in setup
- Write preset config values only if they aren't already set
- Switch DISCOURSE_SMTP_PORT to a preset value
- Change simple_setup to setup
- Use pwd instead of /var/discourse
