+++ launcher
@@ -52,17 +52,17 @@ fi
 usage () {
   echo "Usage: launcher COMMAND CONFIG [--skip-prereqs]"
   echo "Commands:"
-  echo "    start:        Start/initialize a container"
-  echo "    stop:         Stop a running container"
-  echo "    restart:      Restart a container"
-  echo "    destroy:      Stop and remove a container"
-  echo "    enter:        Use nsenter to enter a container"
-  echo "    ssh:          Start a bash shell in a running container"
-  echo "    logs:         Docker logs for container"
-  echo "    bootstrap:    Bootstrap a container for the config based on a template"
-  echo "    rebuild:      Rebuild a container (destroy old, bootstrap, start new)"
-  echo "    cleanup:      Remove all containers that have stopped for > 24 hours"
-  echo "    simple_setup: Configure, bootstrap and start Discourse"
+  echo "    start:      Start/initialize a container"
+  echo "    stop:       Stop a running container"
+  echo "    restart:    Restart a container"
+  echo "    destroy:    Stop and remove a container"
+  echo "    enter:      Use nsenter to enter a container"
+  echo "    ssh:        Start a bash shell in a running container"
+  echo "    logs:       Docker logs for container"
+  echo "    bootstrap:  Bootstrap a container for the config based on a template"
+  echo "    rebuild:    Rebuild a container (destroy old, bootstrap, start new)"
+  echo "    cleanup:    Remove all containers that have stopped for > 24 hours"
+  echo "    setup:      Configure, bootstrap and start Discourse"
   echo
   echo "Options:"
   echo "    --skip-prereqs   Don't check prerequisites or resource requirements"
@@ -171,14 +171,25 @@ check_resources() {
   elif [ "$avail_mem" -lt 1800 ]; then
     total_swap="$(LANG=C free -m | grep ^Swap: | awk '{print $2}')"
     if [ "$total_swap" -lt 1000 ]; then
-      resources="insufficient"
-      echo "WARNING: You must have at least 1GB of swap when running with less"
-      echo "than 2GB of RAM."
-      echo
-      echo "Your system may not work properly, or future upgrades of Discourse may"
-      echo "not complete successfully."
-      echo
-      echo "See https://github.com/discourse/discourse/blob/master/docs/INSTALL-cloud.md#set-up-swap-if-needed"
+      # If we're running setup, configure swap automatically, otherwise, fail and explain
+      if [ "$command" == "setup" ]; then
+        install -o root -g root -m 0600 /dev/null /swapfile
+        dd if=/dev/zero of=/swapfile bs=1k count=2048k
+        mkswap /swapfile
+        swapon /swapfile
+        echo "/swapfile       swap    swap    auto      0       0" | sudo tee -a /etc/fstab
+        sudo sysctl -w vm.swappiness=10
+        echo vm.swappiness = 10 | sudo tee -a /etc/sysctl.conf
+      else
+        resources="insufficient"
+        echo "WARNING: You must have at least 1GB of swap when running with less"
+        echo "than 2GB of RAM."
+        echo
+        echo "Your system may not work properly, or future upgrades of Discourse may"
+        echo "not complete successfully."
+        echo
+        echo "See https://github.com/discourse/discourse/blob/master/docs/INSTALL-cloud.md#set-up-swap-if-needed"
+      fi
     fi
   fi
 
@@ -213,7 +224,7 @@ check_ports() {
   fi
 }
 
-if [ "$opt" != "--skip-prereqs" ] && [ "$command" != "simple_setup" ]; then
+if [ "$opt" != "--skip-prereqs" ] && [ "$command" != "setup" ]; then
   prereqs
 fi
 
@@ -396,7 +407,7 @@ RUBY
   usage
 }
 
-if [ ! -e $config_file ] && [ "$command" != "simple_setup" ]; then
+if [ ! -e $config_file ] && [ "$command" != "setup" ]; then
   echo "Config file was not found, ensure $config_file exists"
   echo ""
   echo "Available configs ( `cd containers && ls -dm *.yml | tr -s '\n' ' ' | awk '{ gsub(/\.yml/, ""); print }'`)"
@@ -417,8 +428,8 @@ fi
 run_ruby_script() {
   ruby_file=$1
   arguments=$2
-  cp containers/"$config".yml ruby_scripts/config.yml  # Add config to the data volume for easy access
-  echo `$docker_path run $user_args --rm -i -a stdout -a stdin -v /var/discourse/ruby_scripts:/ruby_scripts $image ruby /ruby_scripts/"$ruby_file" "$arguments"`
+  cp $config_file ruby_scripts/config.yml  # Add config to the data volume for easy access
+  echo `$docker_path run $user_args --rm -i -a stdout -a stdin -v $(pwd)/ruby_scripts:/ruby_scripts $image ruby /ruby_scripts/"$ruby_file" "$arguments"`
 }
 
 set_existing_container(){
@@ -521,13 +532,12 @@ run_start(){
 
 }
 
-run_simple_setup() {
+run_setup() {
   # Get the config variables by prompting the user for them
   config_variables=(
     DISCOURSE_HOSTNAME
     DISCOURSE_DEVELOPER_EMAILS
     DISCOURSE_SMTP_ADDRESS
-    DISCOURSE_SMTP_PORT
     DISCOURSE_SMTP_USER_NAME
     DISCOURSE_SMTP_PASSWORD
   )
@@ -535,7 +545,6 @@ run_simple_setup() {
     'Domain name'
     "Initial admin and developer email(s)"
     '(email provider) SMTP Address'
-    '(email provider) SMTP Port'
     '(email provider) SMTP Username'
     '(email provider) SMTP Password'
   )
@@ -548,7 +557,7 @@ run_simple_setup() {
   echo
   for index in ${!config_variables[*]}; do # The indexes of the array
     read -p "${config_variables_desc[$index]}: " ${config_variables[$index]}
-    simple_setup_hash+="${config_variables[$index]}, ${!config_variables[$index]},"
+    setup_hash+="${config_variables[$index]}, ${!config_variables[$index]},"
   done
 
   # Copy the sample config if config doesn't exist
@@ -560,7 +569,7 @@ run_simple_setup() {
   prereqs
 
   # Run the simple_installer ruby file with the config variables passed by the user
-  run_ruby_script 'simple_setup.rb' "$simple_setup_hash"
+  run_ruby_script 'setup.rb' "$setup_hash"
   mv ruby_scripts/config.yml containers/"$config".yml  # overwrite the existing config with the updates
 }
 
@@ -619,12 +628,15 @@ run_bootstrap(){
 
 
 case "$command" in
-  simple_setup)
-      run_simple_setup
+  setup)
+      run_setup
       echo "Bootstrapping app"
       echo "This will take between 2 and 8 minutes to complete"
       run_bootstrap
-      echo "Successfully bootstrapped, to startup use ./launcher start $config"
+      echo "Successfully bootstrapped, running startup"
+      run_start
+      echo
+      echo "Setup complete! Discourse is configured and running."
       exit 0
       ;;
